+++ examples/plugins/match/js/main.js
@@ -1,0 +1,7 @@
+/**
+ * LMD match example
+ */
+
+var templates = require.match(/Template$/);
+
+console.log(templates);