+++ CHANGELOG.md
@@ -127,4 +127,5 @@
   - `bundles` - basic part
   - `file_protocol` - tweaks for `file://`
   - `lmdjs_configs` - you can write config files in JavaScript. See [demo](http://lmdjs.org/examples/features/lmdjs_configs/), [code](examples/features/lmdjs_configs/)
+  - `match` - enables `require.match(RegExp): Object` requires every matched module name. See [demo](http://lmdjs.org/examples/plugins/match/), [code](examples/plugins/match/)`